// frontend/src/app/api/schema.d.ts
export interface components {
    schemas: {
// ...
        BrandDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            description?: string | null;
            logoUrl?: string | null;
            productsId?: number[] | null;
        };
// ...
        ColorDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            hexCode?: string | null;
        };
// ...
        ProductDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            /** Format: int32 */
            brandId?: number;
            description?: string | null;
            details?: string | null;
            /** Format: double */
            price?: number;
            productCode?: string | null;
            /** Format: int32 */
            stockQuantity?: number;
            status?: string | null;
            /** Format: int32 */
            colorId?: number;
            colorName?: string | null;
            /** Format: int32 */
            sizeId?: number;
            sizeName?: string | null;
            composition?: string | null;
            categoriesId?: number[] | null;
            matchProductsId?: number[] | null;
            imageUrl?: string[] | null;
            discounts?: components["schemas"]["DiscountDto"][] | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}